ramauno

+++ src/service/productsService.js
@@ -15,12 +15,17 @@ class ProductService {
 		return this.repository.findById(id)
 	}
 
-	async addProduct(body) {
+	async addProduct(body, user) {
+		// Verifica si el usuario tiene el rol adecuado o es el propietario
+		if (!(user && (user.role === 'ADMIN' || user.email === body.owner))) {
+			throw new Error('No tienes permisos para agregar este producto')
+		}
 		return this.repository.create({
 			code: body.code,
 			stock: body.stock,
 			title: body.title,
-			price: body.price
+			price: body.price,
+			owner: body.owner || (user ? user.email : null), // Asigna el propietario solo si hay un usuario autenticado
 		}).then(product => {
 			console.log({ product })
 			sendMail('Nuevo producto agregado', JSON.stringify(product))
@@ -28,11 +33,18 @@ class ProductService {
 		})
 	}
 
-	async updateProduct(id, body) {
+	async updateProduct(id, body, user) {
 		const product = await this.getProductById(id)
 
 		if (!product) {
 			throw new Error('Producto no existe')
+		}
+
+		// Verifica si el usuario tiene permisos para actualizar el producto
+		if (
+			!(user && (user.role === 'ADMIN' || user.email === product.owner))
+		) {
+			throw new Error('No tienes permisos para actualizar este producto')
 		}
 
 		const productUpdated = {
@@ -40,7 +52,7 @@ class ProductService {
 			code: body.code || product.code,
 			stock: body.stock || product.stock,
 			title: body.title || product.title,
-			price: body.price || product.email
+			price: body.price || product.price,
 		}
 
 		await this.repository.updateOne({ _id: id }, productUpdated)
@@ -48,11 +60,18 @@ class ProductService {
 		return productUpdated
 	}
 
-	async deleteProduct(id) {
+	async deleteProduct(id, user) {
 		const product = await this.repository.findById(id)
 
 		if (!product) {
 			throw new Error('Producto no existe')
+		}
+
+		// Verifica si el usuario tiene permisos para eliminar el producto
+		if (
+			!(user && (user.role === 'ADMIN' || (user.role === 'PREMIUM' && product.owner === user.email)))
+		) {
+			throw new Error('No tienes permisos para eliminar este producto')
 		}
 
 		await this.repository.deleteOne({ _id: id })
